requirements.txt
# Email Response Agent - Requirements

# Gmail API
google-auth>=2.48.0
google-auth-oauthlib>=1.2.4
google-auth-httplib2>=0.3.0
google-api-python-client>=2.190.0

# OpenAI API
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0
